+++ lib/src/main/java/provider/SequencesProvider.java
@@ -1,5 +1,5 @@
 package provider;
 
 public interface SequencesProvider {
-    float[] getSequenceByName(String sequenceName);
+    double[] getSequenceByName(String sequenceName);
 }
